+++ frontend/src/container/CreateAlertV2/EvaluationSettings/styles.scss
